添加 （包括css样式）增加学生页面，查询学生页面，删除页面 以及添加部分数据库验证

+++ StudentInfo/views.py
@@ -1,3 +1,4 @@
+from django.http import HttpResponse
 from django.shortcuts import render
 from StudentInfo.models import Student, OneClass
 
@@ -15,22 +16,26 @@ def add(request):
     type1 = request.GET.get('type1')
     # type 为1表示执行数据库操作，否则只是跳转页面
     if type1 == '1':
-        Student.objects.create(name=name)  # 在student 表里新建一条记录
+        Student.objects.create(name=name, age=age, learn=learn)  # 在student 表里新建一条记录
+        # return render(request, 'add.html',)
     return render(request, 'add.html')
 
 
 def delete(request):
     name = request.GET.get('name')
-    # age = request.GET.get('age')
-    # learn = request.GET.get('learn')
     type1 = request.GET.get('type1')
-    # type 为1表示执行数据库操作，否则只是跳转页面
+    # type 为 1 表示执行数据库查询操作
     if type1 == '1':
-        # 在student 表里新建一条记录
+        # 查询所在记录
+        try:
+            stu: Student = Student.objects.filter(name=name).get()
+        except Exception:
+            return HttpResponse("无此记录")
+        return render(request, 'delete.html', {'stu': stu})
+    # type 为 2 表示执行数据库确认删除操作
+    if type1 == '2':
         Student.objects.filter(name=name).delete()
-    return render(request, 'delete.html', {'stu': "删除成功"})
-
-    # rname =
+        return render(request, 'delete.html', {'stu': '删除成功'})
     return render(request, 'delete.html')
